% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cpu_device}
\alias{cpu_device}
\title{CPU device}
\usage{
cpu_device(...)
}
\arguments{
\item{...}{arguments passed to the Julia function.}
}
\value{
a Julia device object
}
\description{
A CPU compute device from \href{https://github.com/LuxDL/MLDataDevices.jl}{MLDataDevices.jl} for the \code{device} argument of \code{\link[=train]{train()}}.
}
\seealso{
\code{\link[=gpu_device]{gpu_device()}}, \code{\link[=reactant_device]{reactant_device()}}, \code{\link[=train]{train()}}
}
