% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{gpu_device}
\alias{gpu_device}
\title{GPU device}
\usage{
gpu_device(...)
}
\arguments{
\item{...}{arguments passed to the Julia function (e.g. \code{force}).}
}
\value{
a Julia device object
}
\description{
A GPU compute device from \href{https://github.com/LuxDL/MLDataDevices.jl}{MLDataDevices.jl} for the \code{device} argument of \code{\link[=train]{train()}}.
}
\seealso{
\code{\link[=cpu_device]{cpu_device()}}, \code{\link[=reactant_device]{reactant_device()}}, \code{\link[=train]{train()}}
}
