% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{infer}
\alias{infer}
\title{infer}
\usage{
infer(estimator, Z, ...)
}
\arguments{
\item{estimator}{a neural estimator}

\item{Z}{data in a format amenable to the neural-network architecture of \code{estimator}}

\item{...}{additional keyword arguments passed to the Julia version of \href{https://msainsburydale.github.io/NeuralEstimators.jl/dev/API/inference#Inference-with-observed-data}{\code{infer()}} (and hence to \code{\link[=estimate]{estimate()}} or \code{\link[=sampleposterior]{sampleposterior()}})}
}
\value{
For neural Bayes estimators, a matrix of point estimates (see \code{\link[=estimate]{estimate()}}). For posterior or ratio estimators, a \eqn{d \times N \times K}{d x N x K} array of posterior samples (see \code{\link[=sampleposterior]{sampleposterior()}}).
}
\description{
Unified inference interface that dispatches to \code{\link[=estimate]{estimate()}} for neural Bayes estimators (e.g., \code{PointEstimator}) and \code{\link[=sampleposterior]{sampleposterior()}} for posterior or ratio estimators.
}
\seealso{
\code{\link[=estimate]{estimate()}}, \code{\link[=sampleposterior]{sampleposterior()}}
}
