% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{loadbestweights}
\alias{loadbestweights}
\title{load the weights of the best neural estimator}
\usage{
loadbestweights(estimator, path)
}
\arguments{
\item{estimator}{the neural point estimator that we wish to load weights into}

\item{path}{absolute path to the folder containing the saved neural-network weights, saved as \code{best_network.bson}}
}
\description{
load the weights of the best neural estimator
}
