% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{loadweights}
\alias{loadweights}
\title{load the weights of a neural estimator}
\usage{
loadweights(estimator, filename)
}
\arguments{
\item{estimator}{the neural estimator that we wish to load weights into}

\item{filename}{file (including absolute path) of the neural-network weights saved as a \code{bson} file}
}
\description{
load the weights of a neural estimator
}
