% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mle}
\alias{mle}
\title{mle}
\usage{
mle(estimator, Z, theta_grid = NULL, use_gpu = TRUE)
}
\arguments{
\item{estimator}{a neural likelihood or likelihood-to-evidence-ratio estimator}

\item{Z}{data; it's format should be amenable to the architecture of \code{estimator}}

\item{theta_grid}{a (fine) gridding of the parameter space, given as a matrix with p rows, where p is the number of parameters in the model}

\item{use_gpu}{a boolean indicating whether to use the GPU if it is available (default true)}
}
\value{
a p × K matrix of maximum-likelihood estimates, where p is the number of parameters in the statistical model and K is the number of data sets provided in \code{Z}
}
\description{
Maximum likelihood estimation
}
