% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralEstimators.R
\docType{package}
\name{NeuralEstimators-package}
\alias{NeuralEstimators}
\alias{NeuralEstimators-package}
\title{NeuralEstimators: Likelihood-Free Parameter Estimation using Neural Networks}
\description{
An 'R' interface to the 'Julia' package 'NeuralEstimators.jl'. The package facilitates the user-friendly development of neural point estimators, which are neural networks that map data to a point summary of the posterior distribution. These estimators are likelihood-free and amortised, in the sense that, after an initial setup cost, inference from observed data can be made in a fraction of the time required by conventional approaches; see Sainsbury-Dale, Zammit-Mangion, and Huser (2024) \doi{10.1080/00031305.2023.2249522} for further details and an accessible introduction. The package also enables the construction of neural networks that approximate the likelihood-to-evidence ratio in an amortised manner, allowing one to perform inference based on the likelihood function or the entire posterior distribution; see Zammit-Mangion, Sainsbury-Dale, and Huser (2024, Sec. 5.2) \doi{10.48550/arXiv.2404.12484}, and the references therein. The package accommodates any model for which simulation is feasible by allowing the user to implicitly define their model through simulated data.
}
\author{
\strong{Maintainer}: Matthew Sainsbury-Dale \email{msainsburydale@gmail.com}

}
