% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotrisk}
\alias{plotrisk}
\title{Plot the risk function (with respect to a given loss function) versus the sample size, m.}
\usage{
plotrisk(df, parameter_labels = NULL, loss = function(x, y) abs(x - y))
}
\arguments{
\item{df}{a long form data frame containing fields \code{estimator},
\code{parameter}, \code{estimate}, \code{truth}, and \code{m}.}

\item{parameter_labels}{a named vector containing parameter labels used in the plot.}

\item{loss}{the loss function; defaults to absolute-error loss.}
}
\value{
a \code{'ggplot'} object showing the risk function versus the sample size, facetted by parameter.
}
\description{
Plot the risk function (with respect to a given loss function) versus the sample size, m.
}
\examples{
# Generate toy data. Two estimators for a single parameter model,
# sample sizes m = 1, 5, 10, 15, 25, and 30, and
# 50 estimates for each combination of estimator and sample size:
m         <- rep(rep(c(1, 5, 10, 15, 20, 25, 30), each = 50), times = 2)
Z         <- lapply(m, rnorm)
estimate  <- sapply(Z, mean)
df <- data.frame(
  estimator = c("Estimator 1", "Estimator 2"),
  parameter = "mu", m = m, estimate = estimate, truth = 0
)

# Plot the risk function
plotrisk(df)
plotrisk(df, loss = function(x, y) (x-y)^2)
}
\seealso{
\code{\link{plotdistribution}}
}
