% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotruntime}
\alias{plotruntime}
\title{Plot the run time versus the sample size, m.}
\usage{
plotruntime(df)
}
\arguments{
\item{df}{a data frame containing fields \code{estimator}, \code{time}, and \code{m}.}
}
\value{
a \code{'ggplot'} object showing the run time versus the sample size.
}
\description{
Plot the run time versus the sample size, m.
}
\examples{
m  <- rep(c(1, 5, 10, 15, 20, 25, 30), each = 2)
df <- data.frame(
  estimator = c("Estimator 1", "Estimator 2"),
  m = m,
  time = 1 / m + 0.1 * runif(length(m))
)
plotruntime(df)
}
\seealso{
\code{\link{plotdistribution}}, \code{\link{plotrisk}}
}
