% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plottrainingrisk}
\alias{plottrainingrisk}
\title{Plots the evolution of the risk function during training.}
\usage{
plottrainingrisk(path, excluded_runs = NULL)
}
\arguments{
\item{path}{a path containing one or more folders titled "runs_x", each of
which corresponds to a training run of a neural estimator and, hence, each
contains a .csv file called 'loss_per_epoch.csv'.}

\item{excluded_runs}{folders to exclude (e.g., "runs_N1" to exclude the folder "runs_N1").}
}
\value{
a \code{'ggplot'} showing the evolution of the risk function during training for each neural estimator.
}
\description{
Plots the evolution of the risk function during training.
}
\seealso{
\code{\link{plotrisk}}, \code{\link{plotdistribution}}, \code{\link{plotruntime}}
}
