% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{reactant_device}
\alias{reactant_device}
\title{Reactant/XLA device}
\usage{
reactant_device(...)
}
\arguments{
\item{...}{arguments passed to the Julia function (e.g. \code{force}).}
}
\value{
a Julia device object
}
\description{
A Reactant/XLA compute device from \href{https://github.com/LuxDL/MLDataDevices.jl}{MLDataDevices.jl} for the \code{device} argument of \code{\link[=train]{train()}}. Requires Lux.
}
\seealso{
\code{\link[=cpu_device]{cpu_device()}}, \code{\link[=gpu_device]{gpu_device()}}, \code{\link[=train]{train()}}
}
