% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{repeatparameters}
\alias{repeatparameters}
\title{repeat parameters in the form expected by NeuralEstimators.jl}
\usage{
repeatparameters(parameters, num_rep)
}
\description{
repeat parameters in the form expected by NeuralEstimators.jl
}
