% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{risk}
\alias{risk}
\title{computes a Monte Carlo approximation of the Bayes risk}
\usage{
risk(
  df,
  loss = function(x, y) abs(x - y),
  average_over_parameters = TRUE,
  average_over_sample_sizes = TRUE
)
}
\arguments{
\item{df}{the \code{estimates} dataframe returned by a call to \code{assess()}}

\item{loss}{a binary operator defining the loss function (default absolute-error loss)}

\item{average_over_parameters}{if \code{TRUE} (default), the loss is averaged over all parameters; otherwise, the loss is averaged over each parameter separately}

\item{average_over_sample_sizes}{if \code{TRUE} (default), the loss is averaged over all sample sizes (the column \code{m} in \code{df}); otherwise, the loss is averaged over each sample size separately}
}
\value{
a dataframe giving the estimated risk and an estimate of its standard deviation
}
\description{
computes a Monte Carlo approximation of the Bayes risk
}
\seealso{
\code{\link[=assess]{assess()}}
}
