% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sampleposterior}
\alias{sampleposterior}
\title{sampleposterior}
\usage{
sampleposterior(estimator, Z, N = 1000, ...)
}
\arguments{
\item{estimator}{a neural posterior or likelihood-to-evidence-ratio estimator}

\item{Z}{data in a format amenable to the neural-network architecture of \code{estimator}}

\item{N}{number of approximate posterior samples to draw}

\item{...}{additional keyword arguments passed to the Julia version of \href{https://msainsburydale.github.io/NeuralEstimators.jl/dev/API/inference#Inference-with-observed-data}{\code{sampleposterior()}}}
}
\value{
a \eqn{d \times N \times K}{d x N x K} array of posterior samples, where \eqn{d} is the dimension of the parameter vector, \code{N} is the number of draws, and \eqn{K} is the number of independent data sets in \code{Z} (so a single data set yields a \eqn{d \times N \times 1}{d x N x 1} array)
}
\description{
Samples from the approximate posterior distribution given data \code{Z}.
}
\seealso{
\code{\link[=infer]{infer()}} for a unified inference interface, and \code{\link[=estimate]{estimate()}} for making inference with neural Bayes estimators
}
