% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bias}
\alias{bias}
\title{computes a Monte Carlo approximation of an estimator's bias}
\usage{
bias(assessment, ...)
}
\arguments{
\item{assessment}{an object returned by \code{assess()} (or the \code{estimates} data frame of this object)}

\item{...}{optional arguments inherited from \code{risk} (excluding the argument \code{loss})}
}
\value{
a dataframe giving the estimated bias
}
\description{
computes a Monte Carlo approximation of an estimator's bias
}
\seealso{
\code{\link[=assess]{assess()}}, \code{\link[=risk]{risk()}}, \code{\link[=rmse]{rmse()}}
}
