% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{loadstate}
\alias{loadstate}
\title{load a saved state of a neural estimator}
\usage{
loadstate(estimator, filename)
}
\arguments{
\item{estimator}{the neural estimator that we wish to load the state into}

\item{filename}{file (including absolute path) of the neural-network state in a \code{bson} file}
}
\value{
\code{estimator} updated with the saved state
}
\description{
Load a saved state of a neural estimator, including both weights and additional model parameters.
}
