% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{loadweights}
\alias{loadweights}
\title{load a collection of saved weights of a neural estimator}
\usage{
loadweights(estimator, filename)
}
\arguments{
\item{estimator}{the neural estimator that we wish to load weights into}

\item{filename}{file (including absolute path) of the neural-network weights saved as a \code{bson} file}
}
\value{
\code{estimator} updated with the saved weights
}
\description{
This function is deprecated and will be removed in a future version.
Please use \code{\link[=loadstate]{loadstate()}} instead, which provides more comprehensive functionality.
}
\seealso{
\code{\link[=loadstate]{loadstate()}}
}
