% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{logratio}
\alias{logratio}
\title{logratio}
\usage{
logratio(estimator, Z, grid, batchsize = 32, ...)
}
\arguments{
\item{estimator}{a neural ratio estimator}

\item{Z}{data in a format amenable to the neural-network architecture of \code{estimator}}

\item{grid}{matrix of parameter values, where each column is a parameter configuration}

\item{batchsize}{the batch size}

\item{...}{additional keyword arguments passed to the Julia version of \href{https://msainsburydale.github.io/NeuralEstimators.jl/dev/API/inference#Inference-with-observed-data}{\code{logratio()}}}
}
\value{
A matrix of log ratios with one row per data set and one column per grid point
}
\description{
Apply a neural ratio estimator to data
}
\seealso{
\code{\link[=sampleposterior]{sampleposterior()}} for making posterior inferences
}
