% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rmse}
\alias{rmse}
\title{computes a Monte Carlo approximation of an estimator's root-mean-square error (RMSE)}
\usage{
rmse(assessment, ...)
}
\arguments{
\item{assessment}{an object returned by \code{assess()} (or the \code{estimates} data frame of this object)}

\item{...}{optional arguments inherited from \code{risk} (excluding the argument \code{loss})}
}
\value{
a dataframe giving the estimated RMSE
}
\description{
computes a Monte Carlo approximation of an estimator's root-mean-square error (RMSE)
}
\seealso{
\code{\link[=assess]{assess()}}, \code{\link[=bias]{bias()}}, \code{\link[=risk]{risk()}}
}
