% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sampleposterior}
\alias{sampleposterior}
\title{sampleposterior}
\usage{
sampleposterior(
  estimator,
  Z,
  N = 1000,
  prior = NULL,
  theta_grid = NULL,
  use_gpu = TRUE
)
}
\arguments{
\item{estimator}{a neural likelihood, posterior, or likelihood-to-evidence-ratio estimator}

\item{Z}{data; it's format should be amenable to the architecture of \code{estimator}}

\item{N}{number of samples (default 1000)}

\item{prior}{the prior (default uniform), specified as a Julia or R function}

\item{theta_grid}{a (fine) gridding of the parameter space, given as a matrix with p rows, where p is the number of parameters in the model}

\item{use_gpu}{a boolean indicating whether to use the GPU if it is available (default true)}
}
\value{
a p × \code{N} matrix of posterior samples, where p is the number of parameters in the model. If multiple data sets are given in \code{Z}, a list of posterior samples will be returned
}
\description{
Sample from the approximate posterior distribution
}
