% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{savestate}
\alias{savestate}
\title{save the state of a neural estimator}
\usage{
savestate(estimator, filename)
}
\arguments{
\item{estimator}{the neural estimator that we wish to save}

\item{filename}{file in which to save the neural-network state as a \code{bson} file}
}
\value{
No return value, called for side effects
}
\description{
save the state of a neural estimator
}
