% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{train}
\alias{train}
\title{train a neural estimator}
\usage{
train(
  estimator,
  theta_train,
  theta_val,
  Z_train,
  Z_val,
  M = NULL,
  epochs = 100L,
  batchsize = 256L,
  savepath = "runs/",
  stopping_epochs = 10L,
  use_gpu = TRUE,
  verbose = TRUE
)
}
\description{
train a neural estimator
}
